Refactor Watch Controller and Watch Service

+++ TrackMyWristAPI/Controllers/WatchController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using TrackMyWristAPI.Dtos.Watch;
+using TrackMyWristAPI.Services.OwnershipService;
 using TrackMyWristAPI.Services.WatchService;
 
 namespace TrackMyWristAPI.Controllers
@@ -17,9 +18,11 @@ namespace TrackMyWristAPI.Controllers
     public class WatchController : ControllerBase
     {
         private readonly IWatchService _watchService;
+        private readonly IOwnershipService _ownershipService;
 
-        public WatchController(IWatchService watchService)
+        public WatchController(IWatchService watchService, IOwnershipService ownershipService)
         {
+            _ownershipService = ownershipService;
             _watchService = watchService;
         }
 
@@ -34,14 +37,15 @@ namespace TrackMyWristAPI.Controllers
         [HttpGet("{id:int}", Name = "getWatchById")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<ActionResult<GetWatchDto>> GetWatchById(int id)
         {
-            var existingWatch = await _watchService.GetWatchById(id);
-            if (existingWatch == null)
+            if (!await _ownershipService.WatchBelongsToUser(id))
             {
-                return NotFound();
+                return Forbid();
             }
+            var existingWatch = await _watchService.GetWatchById(id);
             return Ok(existingWatch);
         }
 
@@ -59,28 +63,30 @@ namespace TrackMyWristAPI.Controllers
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<ActionResult<GetWatchDto>> UpdateWatch(int id, [FromBody] UpdateWatchDto watch)
         {
-            var updatedWatch = await _watchService.UpdateWatch(id, watch);
-            if (updatedWatch == null)
+            if (!await _ownershipService.WatchBelongsToUser(id))
             {
-                return NotFound();
+                return Forbid();
             }
+            var updatedWatch = await _watchService.UpdateWatch(id, watch);
             return Ok(updatedWatch);
         }
 
         [HttpDelete("{id:int}")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<ActionResult<GetWatchDto>> DeleteWatch(int id)
         {
-            var deletedWatch = await _watchService.DeleteWatch(id);
-            if (deletedWatch == null)
+            if (!await _ownershipService.WatchBelongsToUser(id))
             {
-                return NotFound();
+                return Forbid();
             }
+            await _watchService.DeleteWatch(id);
             return NoContent();
         }
     }
